Upgrade laravel and laravel-mix

+++ database/migrations/2016_12_11_153312_create_followers_table.php
@@ -14,9 +14,9 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('follow_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('follow_id');
             $table->timestamps();
         });
     }
